Created specific service for multer code
